+++ src/components/nivo/legend-symbol/index.js
