+++ src/components/nivo/shared/constants/dimensions.js
