+++ src/index.js
@@ -1,7 +1,7 @@
-export { default as BarChart } from './components/bar-chart'
-export { default as LineChart } from './components/line-chart'
-export { default as ScatterChart } from './components/scatter-chart'
-export { default as PieChart } from './components/pie-chart'
+export { default as BarChart } from './components/nivo/bar-chart'
+export { default as LineChart } from './components/nivo/line-chart'
+export { default as ScatterChart } from './components/nivo/scatter-chart'
+export { default as PieChart } from './components/nivo/pie-chart'
 export { default as Cluster } from './components/d3/cluster'
 export { default as GaugeBar } from './components/d3/gauge-bar'
 export { default as GaugeArc } from './components/d3/gauge-arc'
