+++ stories/nivo/color.stories.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import BarChart from '../../src/components/bar-chart/'
-import PieChart from '../../src/components/pie-chart'
-import LineChart from '../../src/components/line-chart'
-import ScatterChart from '../../src/components/scatter-chart'
+import BarChart from '../../src/components/nivo/bar-chart/'
+import PieChart from '../../src/components/nivo/pie-chart'
+import LineChart from '../../src/components/nivo/line-chart'
+import ScatterChart from '../../src/components/nivo/scatter-chart'
 import barChartData from '../data/others/bar-chart-data'
 import lineChartData from '../data/others/line-chart-data'
 import pieChartData from '../data/others/pie-chart-data'
