+++ stories/plotly/plotly-bar.stories.js
@@ -15,12 +15,12 @@ const data = [
     x: keyX,
     y: keyY,
     type: 'bar',
-  }
+  },
 ]
 export const Default = () => <Bar data={data} />
 
 export const Responsive = () => <div style={{
-  width: '100%', height: '100vh'
+  width: '100%', height: '100vh',
 }}><ParentSize>{({ width, height }) => {
-  return (<Bar data={data} layout={{ width: width, height: height }} config={{ displayModeBar: false }} />)
-}}</ParentSize></div>
+    return (<Bar data={data} layout={{ width: width, height: height }} config={{ displayModeBar: false }} />)
+  }}</ParentSize></div>
